Move database helper into Utils
