+++ lib/snowflakes/cli/commands/db/create_migration.rb
@@ -1,6 +1,7 @@
 require "hanami/cli"
 require "snowflakes/cli/command"
 require_relative "structure/dump"
+require_relative "utils/database"
 
 module Snowflakes
   class CLI < Hanami::CLI
@@ -12,12 +13,18 @@ module Snowflakes
           argument :name, desc: "Migration file name"
 
           def call(name:, **)
-            migrator = application.database.migrator
+            migrator = database.migrator
             version = migrator.generate_version
 
             measure "migration #{version}_#{name} created" do
               migrator.create_file(name, version)
             end
+          end
+
+          private
+
+          def database
+            @database ||= Utils::Database.for_application(application)
           end
         end
       end
