+++ lib/snowflakes/cli/commands/db/migrate.rb
@@ -1,6 +1,7 @@
 require "hanami/cli"
 require "snowflakes/cli/command"
 require_relative "structure/dump"
+require_relative "utils/database"
 
 module Snowflakes
   class CLI < Hanami::CLI
@@ -12,8 +13,6 @@ module Snowflakes
           option :target, desc: "Target migration number", aliases: ["-t"]
 
           def call(target: nil, **)
-            database = application.database
-
             measure "database #{database.name} migrated" do
               if target
                 database.gateway.run_migrations(target: target)
@@ -23,6 +22,12 @@ module Snowflakes
             end
 
             run_command Structure::Dump
+          end
+
+          private
+
+          def database
+            @database ||= Utils::Database.for_application(application)
           end
         end
       end
